Fixed middleware error

+++ part4/bloglist/controllers/blogs.js
@@ -42,6 +42,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const body = request.body
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  
+  const blog = await Blog.findById(decodedToken.id)
+  
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
